+++ CHANGES.md
@@ -6,6 +6,7 @@ For more information, check the README.md in <http://github.com/i3visio/osrframe
 0.13.0a, 2016/07/09 -- Fixed issues #146 y #147.
 - Fixed issue #147. Letting the user create new entify regexps in its home folder.
 - Fixed issue #146. Letting the user create new usufy wrappers in its home folder.
+- Addressed issue #149. Bookmarky.com seems not to be working (moved to pending).
 - Modified the configuration file.
 - Removed old wrappers which had been removed.
 - Added a pending folder with wrappers which need to be fixed. They will not be imported.